Implement SSR in Switch
