+++ testing/tests/DevExpress.serverSide/switch.markup.tests.js
@@ -1,0 +1,3 @@
+"use strict";
+
+require("../DevExpress.ui.widgets.editors/switch.markup.tests.js");
